refactor(analytics): Convert form trackers to functional components and update tests
