+++ packages/alpha-app/src/views/common/base.njk
