+++ packages/frontend-analytics/src/analytics/formChangeTracker/formChangeTracker.test.ts
@@ -1,5 +1,5 @@
 import { beforeEach, describe, expect, jest, test } from "@jest/globals";
-import { FormChangeTracker } from "./formChangeTracker";
+import * as formChangeTracker from "./formChangeTracker";
 import * as pushToDataLayer from "../../utils/pushToDataLayer";
 
 function createForm() {
@@ -18,8 +18,11 @@ function createForm() {
 }
 
 describe("FormChangeTracker", () => {
-  let newInstance: FormChangeTracker;
-  let action: MouseEvent;
+  const action: MouseEvent = new MouseEvent("click", {
+    view: window,
+    bubbles: true,
+    cancelable: true,
+  });
 
   beforeEach(() => {
     jest.clearAllMocks();
@@ -27,30 +30,19 @@ describe("FormChangeTracker", () => {
     window.DI = { analyticsGa4: { cookie: { consent: true } } };
 
     jest.spyOn(pushToDataLayer, "pushToDataLayer");
-    jest.spyOn(FormChangeTracker.prototype, "initialiseEventListener");
-    jest.spyOn(FormChangeTracker.prototype, "trackFormChange");
-
-    const enableFormChangeTracking = true;
-    newInstance = new FormChangeTracker(enableFormChangeTracking);
-    action = new MouseEvent("click", {
-      view: window,
-      bubbles: true,
-      cancelable: true,
-    });
+    jest.spyOn(formChangeTracker, "FormChangeTracker");
   });
 
-  test("form change should be deactivated, if flag is set to false", () => {
-    const instance = new FormChangeTracker(false);
-    expect(instance.trackFormChange).not.toBeCalled();
+  test("should not call pushToDataLayer if enableFormChangeTracking is false", () => {
+    formChangeTracker.FormChangeTracker(false);
+    expect(pushToDataLayer.pushToDataLayer).not.toBeCalled();
   });
 
-  test("new instance should call initialiseEventListener", () => {
-    expect(newInstance.initialiseEventListener).toBeCalled();
-  });
-
-  test("pushToDataLayer is called", () => {
+  test("should call pushToDataLayer if enableFormChangeTracking is true", () => {
+    formChangeTracker.FormChangeTracker(true);
     const { changeLink } = createForm();
     changeLink.dispatchEvent(action);
+
     expect(pushToDataLayer.pushToDataLayer).toBeCalledWith({
       event: "event_data",
       event_data: {
@@ -71,14 +63,14 @@ describe("FormChangeTracker", () => {
     });
   });
 
-  test("trackFormChange should return false if not cookie consent", () => {
+  test("should not call pushToDataLayer if cookie consent is false", () => {
     window.DI.analyticsGa4.cookie.consent = false;
     const { changeLink } = createForm();
     changeLink.dispatchEvent(action);
     expect(pushToDataLayer.pushToDataLayer).not.toHaveBeenCalled();
   });
 
-  test("trackFormChange should return false if not a change link", () => {
+  test("should not call pushToDataLayer if event is not from a change link", () => {
     const { form } = createForm();
 
     const href = document.createElement("div");
@@ -90,66 +82,16 @@ describe("FormChangeTracker", () => {
     expect(pushToDataLayer.pushToDataLayer).not.toHaveBeenCalled();
   });
 
-  test("trackFormChange should return true if a Change link", () => {
+  test("should call pushToDataLayer if event is from a change link", () => {
     const { changeLink } = createForm();
     changeLink.dispatchEvent(action);
     expect(pushToDataLayer.pushToDataLayer).toHaveBeenCalled();
   });
 
-  test("trackFormChange should return false if it is a Lang Toggle link", () => {
+  test("should not call pushToDataLayer if event is not from a Lang Toggle link", () => {
     const { changeLink } = createForm();
     changeLink.setAttribute("hreflang", "en");
     changeLink.dispatchEvent(action);
     expect(pushToDataLayer.pushToDataLayer).not.toHaveBeenCalled();
-  });
-
-  test("should return 'undefined' if parent element does not exist", () => {
-    document.body.innerHTML = `<a id="change_link" href="http://localhost?edit=true">Change</a>`;
-    const href = document.getElementById("change_link") as HTMLAnchorElement;
-    expect(FormChangeTracker.getSection(href)).toBe("undefined");
-  });
-
-  test("should return text content of sibling with class 'govuk-summary-list__key'", () => {
-    document.body.innerHTML = `
-    <div class="govuk-summary-list__key">Expected Section</div>
-    <div>
-      <a id="change_link">Change</a>
-    </div>
-  `;
-    const href = document.getElementById("change_link") as HTMLAnchorElement;
-    expect(FormChangeTracker.getSection(href)).toBe("Expected Section");
-  });
-
-  test("should check and return text content of parent element if no matching sibling found", () => {
-    document.body.innerHTML = `  
-    <div>
-    Postcode
-      <a id="change_link">Change</a> 
-    </div>
-  `;
-    const href = document.getElementById("change_link") as HTMLAnchorElement;
-    expect(FormChangeTracker.getSection(href)).toBe("Postcode");
-  });
-
-  test("should return 'undefined' if no matching sibling found and parent has no text content", () => {
-    document.body.innerHTML = `
-      <div>
-        <a id="change_link">Change</a>
-      </div>
-    `;
-    const href = document.getElementById("change_link") as HTMLAnchorElement;
-
-    expect(FormChangeTracker.getSection(href)).toBe("undefined");
-  });
-
-  test("should return 'undefined' if summary list key has no text content", () => {
-    document.body.innerHTML = `
-    <div class="govuk-summary-list__key"></div>
-    <div>
-      <a id="change_link">Change</a>
-    </div>
-  `;
-    const href = document.getElementById("change_link") as HTMLAnchorElement;
-    expect(FormChangeTracker.getSection(href)).toBe("undefined");
   });
 });
