+++ packages/frontend-analytics/src/analytics/formChangeTracker/formChangeTracker.ts
@@ -1,44 +1,30 @@
 import logger from "loglevel";
 import { validateParameter } from "../../utils/validateParameter";
-import { FormTracker } from "../formTracker/formTracker";
-import { FormEventInterface } from "../formTracker/formTracker.interface";
+import * as FormTrackerUtils from "../../utils/formUtils/formTrackerUtils";
+import { FormEventInterface } from "../../utils/formUtils/formTracker.interface";
 import {
   getDomain,
   getDomainPath,
   isChangeLink,
 } from "../../utils/dataScrapers";
 import { pushToDataLayer } from "../../utils/pushToDataLayer";
+import { getSection } from "./utils/formChangeTracker.utils";
 
-export class FormChangeTracker extends FormTracker {
-  eventName: string = "form_change_response";
-  eventType: string = "event_data";
-  enableFormChangeTracking: boolean;
-
-  constructor(enableFormChangeTracking: boolean) {
-    super();
-    this.enableFormChangeTracking = enableFormChangeTracking;
-    this.initialiseEventListener();
-  }
-
-  initialiseEventListener() {
-    document.addEventListener("click", this.trackFormChange.bind(this), false);
-  }
+export const FormChangeTracker = (enableFormChangeTracking: boolean) => {
+  const eventName: string = "form_change_response";
+  const eventType: string = "event_data";
 
   /**
    * Tracks changes in a form and sends data to the analytics platform.
    *
    * @return {boolean} Returns true if the form change tracking is successful, otherwise false.
    */
-  trackFormChange(event: Event): boolean {
-    if (!window.DI.analyticsGa4.cookie.consent) {
+  const trackFormChange = (event: Event): boolean => {
+    if (!window.DI.analyticsGa4.cookie.consent || !enableFormChangeTracking) {
       return false;
     }
 
-    if (!this.enableFormChangeTracking) {
-      return false;
-    }
-
-    const form = FormTracker.getFormElement();
+    const form = FormTrackerUtils.getFormElement();
 
     if (!form) {
       return false;
@@ -56,13 +42,13 @@ export class FormChangeTracker extends FormTracker {
     }
 
     const formChangeTrackerEvent: FormEventInterface = {
-      event: this.eventType,
+      event: eventType,
       event_data: {
-        event_name: this.eventName,
+        event_name: eventName,
         type: "undefined",
         url: validateParameter(element.href, 500),
         text: "change", // put static value. Waiting final documentation on form change tracker,
-        section: validateParameter(FormChangeTracker.getSection(element), 100),
+        section: validateParameter(getSection(element), 100),
         action: "change response",
         external: "false",
         link_domain: getDomain(element.href),
@@ -81,32 +67,7 @@ export class FormChangeTracker extends FormTracker {
       logger.error("Error in trackFormChange", err);
       return false;
     }
-  }
+  };
 
-  static getSection(element: HTMLElement): string {
-    const { parentElement } = element;
-
-    // Ensure the parent element exists
-    if (!parentElement) {
-      return "undefined";
-    }
-
-    let parentSiblingElement = parentElement?.previousElementSibling;
-    while (parentSiblingElement) {
-      if (parentSiblingElement.classList.contains("govuk-summary-list__key")) {
-        const sectionValue =
-          parentSiblingElement.textContent?.trim() || "undefined";
-        return sectionValue;
-      }
-      parentSiblingElement = parentSiblingElement.previousElementSibling;
-    }
-    // If no matching sibling is found, check the parent element
-    let parentTextContent = "";
-    parentElement.childNodes.forEach((node) => {
-      if (node.nodeType === Node.TEXT_NODE) {
-        parentTextContent += node.textContent?.trim() || "";
-      }
-    });
-    return parentTextContent || "undefined";
-  }
-}
+  document.addEventListener("click", trackFormChange);
+};
