+++ packages/frontend-analytics/src/analytics/formResponseTracker/formResponseTracker.ts
@@ -1,48 +1,34 @@
 import logger from "loglevel";
 import { validateParameter } from "../../utils/validateParameter";
-import { FormTracker } from "../formTracker/formTracker";
 import {
   FormEventInterface,
   FormField,
-} from "../formTracker/formTracker.interface";
+} from "../../utils/formUtils/formTracker.interface";
 import { getDomain, getDomainPath } from "../../utils/dataScrapers";
 import { pushToDataLayer } from "../../utils/pushToDataLayer";
+import {
+  combineDateFields,
+  getButtonLabel,
+  getFieldValue,
+  getFormFields,
+  isDateFields,
+  isFormValid,
+  redactPII,
+} from "./utils/formResponseTracker.utils";
+import {
+  getFormElement,
+  getSubmitUrl,
+  getFieldType,
+  getSectionValue,
+} from "../../utils/formUtils/formTrackerUtils";
 
-export class FormResponseTracker extends FormTracker {
-  eventName: string = "form_response";
-  eventType: string = "event_data";
-  isDataSensitive: boolean;
-  isPageDataSensitive: boolean = false;
-  enableFormResponseTracking: boolean;
-
-  /**
-   * Initializes a new instance of the FormResponseTracker class.
-   *  * @param {boolean} isDataSensitive - Flag if data is sensitive
-   *  * @return {void}
-   */
-  constructor(
-    isDataSensitive: boolean,
-    isPageDataSensitive: boolean,
-    enableFormResponseTracking: boolean,
-  ) {
-    super();
-    this.isDataSensitive = isDataSensitive;
-    this.isPageDataSensitive = isPageDataSensitive;
-    this.enableFormResponseTracking = enableFormResponseTracking;
-    this.initialiseEventListener();
-  }
-
-  /**
-   * Initialise the event listener for the document.
-   *
-   */
-  initialiseEventListener(): void {
-    document.addEventListener(
-      "submit",
-      this.trackFormResponse.bind(this),
-      false,
-    );
-  }
+export const FormResponseTracker = (
+  isDataSensitive: boolean,
+  isPageDataSensitive: boolean,
+  enableFormResponseTracking: boolean,
+) => {
+  const eventName: string = "form_response";
+  const eventType: string = "event_data";
 
   /**
    * Tracks the form response and sends the data to the analytics platform.
@@ -50,16 +36,12 @@ export class FormResponseTracker extends FormTracker {
    * @param {SubmitEvent} event - The submit event triggered by the form submission.
    * @return {boolean} Returns true if the form response is successfully tracked, otherwise false.
    */
-  trackFormResponse(event: SubmitEvent): boolean {
-    if (!window.DI.analyticsGa4.cookie.consent) {
+  const trackFormResponse = (event: SubmitEvent): boolean => {
+    if (!window.DI.analyticsGa4.cookie.consent || !enableFormResponseTracking) {
       return false;
     }
 
-    if (!this.enableFormResponseTracking) {
-      return false;
-    }
-
-    const form = FormTracker.getFormElement();
+    const form = getFormElement();
 
     if (!form) {
       return false;
@@ -68,43 +50,37 @@ export class FormResponseTracker extends FormTracker {
     let fields: FormField[] = [];
 
     if (form?.elements) {
-      fields = this.getFormFields(form);
+      fields = getFormFields(form);
     } else {
       return false;
     }
 
-    if (!fields.length) {
-      return false;
-    }
-
-    // check if form is valid
-    if (!FormTracker.isFormValid(fields)) {
+    if (!fields.length || !isFormValid(fields)) {
       return false;
     }
 
     // manage date (day/month/year) fields
-    if (FormTracker.isDateFields(fields)) {
-      fields = FormTracker.combineDateFields(fields);
+    if (isDateFields(fields)) {
+      fields = combineDateFields(fields);
     }
 
-    const submitUrl = FormTracker.getSubmitUrl(form);
+    const submitUrl = getSubmitUrl(form);
 
     try {
       fields.forEach((field) => {
         const formResponseTrackerEvent: FormEventInterface = {
-          event: this.eventType,
+          event: eventType,
           event_data: {
-            event_name: this.eventName,
-            type: validateParameter(this.getFieldType([field]), 100),
+            event_name: eventName,
+            type: validateParameter(getFieldType([field]), 100),
             url: validateParameter(submitUrl, 100),
-            text: this.redactPII(
-              validateParameter(FormTracker.getFieldValue([field]), 100),
+            text: redactPII(
+              validateParameter(getFieldValue([field]), 100),
+              isDataSensitive,
+              isPageDataSensitive
             ),
-            section: validateParameter(FormTracker.getSectionValue(field), 100),
-            action: validateParameter(
-              FormResponseTracker.getButtonLabel(event),
-              100,
-            ),
+            section: validateParameter(getSectionValue(field), 100),
+            action: validateParameter(getButtonLabel(event), 100),
             external: "false",
             link_domain: getDomain(submitUrl),
             "link_path_parts.1": getDomainPath(submitUrl, 0),
@@ -124,29 +100,11 @@ export class FormResponseTracker extends FormTracker {
       logger.error("Error in trackFormResponse", err);
       return false;
     }
-  }
+  };
 
   /**
-   * Retrieves the label of a button from a SubmitEvent.
+   * Initialise the event listener for the document.
    *
-   * @param {SubmitEvent} event - The SubmitEvent object containing the button.
-   * @return {string} The label of the button, or "undefined" if it is not found.
    */
-  static getButtonLabel(event: SubmitEvent): string {
-    return event.submitter?.textContent
-      ? event.submitter.textContent.trim()
-      : "undefined";
-  }
-
-  /**
-   * Redacts field value if data is flagged as sensitive
-   *
-   * @param {string} string - Text field input
-   * @return {string} The text field or undefined
-   */
-  redactPII(parameter: string): string {
-    return this.isPageDataSensitive || this.isDataSensitive
-      ? "undefined"
-      : parameter.trim();
-  }
-}
+  document.addEventListener("submit", trackFormResponse);
+};
