+++ packages/frontend-analytics/src/analytics/pageViewTracker/pageViewTracker.test.ts
@@ -7,9 +7,17 @@ import {
 } from "./pageViewTracker.interface";
 import { PageViewTracker } from "./pageViewTracker";
 import { OptionsInterface } from "../core/core.interface";
-import { FormErrorTracker } from "../formErrorTracker/formErrorTracker";
-import { FormChangeTracker } from "../formChangeTracker/formChangeTracker";
+import * as FormErrorTracker from "../formErrorTracker/formErrorTracker";
 import * as pushToDataLayer from "../../utils/pushToDataLayer";
+import {
+  getLanguage,
+  getLocation,
+  getReferrer,
+  getLoggedInStatus,
+  getFirstPublishedAt,
+  getUpdatedAt,
+  getRelyingParty,
+} from "./utils/pageViewTracker.utils";
 
 window.DI = { analyticsGa4: { cookie: { consent: true } } };
 
@@ -83,12 +91,12 @@ describe("pageViewTracker", () => {
     const dataLayerEvent: PageViewEventInterface = {
       event: instance.eventName,
       page_view: {
-        language: PageViewTracker.getLanguage(),
-        location: PageViewTracker.getLocation(),
+        language: getLanguage(),
+        location: getLocation(),
         organisations: instance.organisations,
         primary_publishing_organisation:
           instance.primary_publishing_organisation,
-        referrer: PageViewTracker.getReferrer(),
+        referrer: getReferrer(),
         status_code: parameters.statusCode.toString(),
         title: parameters.englishPageTitle,
         taxonomy_level1: parameters.taxonomy_level1,
@@ -97,13 +105,11 @@ describe("pageViewTracker", () => {
         taxonomy_level4: parameters.taxonomy_level4,
         taxonomy_level5: parameters.taxonomy_level5,
         content_id: parameters.content_id,
-        logged_in_status: PageViewTracker.getLoggedInStatus(
-          parameters.logged_in_status,
-        ),
+        logged_in_status: getLoggedInStatus(parameters.logged_in_status),
         dynamic: parameters.dynamic.toString(),
-        first_published_at: PageViewTracker.getFirstPublishedAt(),
-        updated_at: PageViewTracker.getUpdatedAt(),
-        relying_party: PageViewTracker.getRelyingParty(),
+        first_published_at: getFirstPublishedAt(),
+        updated_at: getUpdatedAt(),
+        relying_party: getRelyingParty(),
       },
     };
     instance.trackOnPageLoad(parameters);
@@ -115,12 +121,12 @@ describe("pageViewTracker", () => {
     const dataLayerEvent: PageViewEventInterface = {
       event: instance.eventName,
       page_view: {
-        language: PageViewTracker.getLanguage(),
-        location: PageViewTracker.getLocation(),
+        language: getLanguage(),
+        location: getLocation(),
         organisations: instance.organisations,
         primary_publishing_organisation:
           instance.primary_publishing_organisation,
-        referrer: PageViewTracker.getReferrer(),
+        referrer: getReferrer(),
         status_code: parameters.statusCode.toString(),
         title: parameters.englishPageTitle,
         taxonomy_level1: parameters.taxonomy_level1,
@@ -129,65 +135,15 @@ describe("pageViewTracker", () => {
         taxonomy_level4: parameters.taxonomy_level4,
         taxonomy_level5: parameters.taxonomy_level5,
         content_id: parameters.content_id,
-        logged_in_status: PageViewTracker.getLoggedInStatus(
-          parameters.logged_in_status,
-        ),
+        logged_in_status: getLoggedInStatus(parameters.logged_in_status),
         dynamic: parameters.dynamic.toString(),
-        first_published_at: PageViewTracker.getFirstPublishedAt(),
-        updated_at: PageViewTracker.getUpdatedAt(),
-        relying_party: PageViewTracker.getRelyingParty(),
+        first_published_at: getFirstPublishedAt(),
+        updated_at: getUpdatedAt(),
+        relying_party: getRelyingParty(),
       },
     };
     instance.trackOnPageLoad(parameters);
     expect(pushToDataLayer.pushToDataLayer).toBeCalledWith(dataLayerEvent);
-  });
-
-  test("getLoggedInStatus returns the good data if logged in", () => {
-    const status = PageViewTracker.getLoggedInStatus(true);
-    expect(status).toBe("logged in");
-  });
-
-  test("getLoggedInStatus returns the good data if logged out", () => {
-    const status = PageViewTracker.getLoggedInStatus(false);
-    expect(status).toBe("logged out");
-  });
-
-  test("getLoggedInStatus returns the good data if loggedinstatus is undefined", () => {
-    const status = PageViewTracker.getLoggedInStatus(undefined);
-    expect(status).toBe("undefined");
-  });
-
-  test("getRelyingParty returns the good data", () => {
-    const relyingParty = PageViewTracker.getRelyingParty();
-    expect(relyingParty).toBe("localhost");
-  });
-
-  test("getFirstPublishedAt returns undefined if first published-at tag doesn't exists", () => {
-    const firstPublishedAt = PageViewTracker.getFirstPublishedAt();
-    expect(firstPublishedAt).toBe("undefined");
-  });
-
-  test("getFirstPublishedAt returns the good data if first published-at tag exists", () => {
-    const newTag = document.createElement("meta");
-    newTag.setAttribute("name", "govuk:first-published-at");
-    newTag.setAttribute("content", "2022-09-01T00:00:00.000Z");
-    document.head.appendChild(newTag);
-    const firstPublishedAt = PageViewTracker.getFirstPublishedAt();
-    expect(firstPublishedAt).toBe("2022-09-01T00:00:00.000Z");
-  });
-
-  test("getUpdatedAt returns undefined if updated-at tag doesn't exists", () => {
-    const updatedAt = PageViewTracker.getUpdatedAt();
-    expect(updatedAt).toBe("undefined");
-  });
-
-  test("getUpdatedAt returns the good data if updated-at tag exists", () => {
-    const newTag = document.createElement("meta");
-    newTag.setAttribute("name", "govuk:updated-at");
-    newTag.setAttribute("content", "2022-09-02T00:00:00.000Z");
-    document.head.appendChild(newTag);
-    const updatedAt = PageViewTracker.getUpdatedAt();
-    expect(updatedAt).toBe("2022-09-02T00:00:00.000Z");
   });
 });
 
@@ -221,8 +177,6 @@ describe("Cookie Management", () => {
 });
 
 describe("Form Change Tracker Trigger", () => {
-  const spy = jest.spyOn(FormChangeTracker.prototype, "trackFormChange");
-
   test("FormChange tracker is not triggered", () => {
     const instance = new PageViewTracker(
       getOptions({
@@ -231,22 +185,20 @@ describe("Form Change Tracker Trigger", () => {
     );
 
     instance.trackOnPageLoad(getParameters());
-    expect(spy).not.toHaveBeenCalled();
+    expect(pushToDataLayer.pushToDataLayer).not.toHaveBeenCalled();
   });
 });
 
 describe("Form Error Tracker Trigger", () => {
   let instance: PageViewTracker;
-  let formErrorTracker: FormErrorTracker;
 
   beforeEach(() => {
     jest.clearAllMocks();
     window.DI.analyticsGa4.cookie.hasCookie = true;
     window.DI.analyticsGa4.cookie.consent = true;
     jest.spyOn(pushToDataLayer, "pushToDataLayer");
-    jest.spyOn(FormErrorTracker.prototype, "trackFormError");
+    jest.spyOn(FormErrorTracker, "trackFormError");
     instance = new PageViewTracker(getOptions());
-    formErrorTracker = new FormErrorTracker();
   });
 
   test("trackOnPageLoad should called form error function and return false if form error message exists", () => {
@@ -261,7 +213,7 @@ describe("Form Error Tracker Trigger", () => {
       '<p id="organisationType-error" class="govuk-error-message"><span class="govuk-visually-hidden">Error:</span> Select one option</p>';
 
     instance.trackOnPageLoad(getParameters());
-    expect(formErrorTracker.trackFormError).toHaveBeenCalled();
+    expect(FormErrorTracker.trackFormError).toHaveBeenCalled();
   });
   test("FormError tracker is activated even if pageView is disabled", () => {
     document.body.innerHTML =
@@ -273,13 +225,13 @@ describe("Form Error Tracker Trigger", () => {
     );
 
     newInstance.trackOnPageLoad(getParameters());
-    expect(formErrorTracker.trackFormError).toHaveBeenCalled();
+    expect(FormErrorTracker.trackFormError).toHaveBeenCalled();
   });
 
   test("FormError tracker is not triggered", () => {
     document.body.innerHTML = "";
     instance.trackOnPageLoad(getParameters());
-    expect(formErrorTracker.trackFormError).not.toBeCalled();
+    expect(FormErrorTracker.trackFormError).not.toBeCalled();
   });
 
   test("FormError tracker is deactivated", () => {
@@ -328,7 +280,7 @@ describe("Persisting taxonomy level values", () => {
       page_view: {
         content_id: "<e4a3603d-2d3c-4ff1-9b80-d72c1e6b7a58>",
         dynamic: "true",
-        first_published_at: PageViewTracker.getFirstPublishedAt(),
+        first_published_at: getFirstPublishedAt(),
         language: "undefined",
         location: "http://localhost/",
         logged_in_status: "logged in",
@@ -344,7 +296,7 @@ describe("Persisting taxonomy level values", () => {
         taxonomy_level4: "taxo4",
         taxonomy_level5: "taxo5",
         title: "home",
-        updated_at: PageViewTracker.getUpdatedAt(),
+        updated_at: getUpdatedAt(),
       },
     });
     expect(pushToDataLayer.pushToDataLayer).toHaveBeenNthCalledWith(2, {
@@ -352,7 +304,7 @@ describe("Persisting taxonomy level values", () => {
       page_view: {
         content_id: "<e4a3603d-2d3c-4ff1-9b80-d72c1e6b7a58>",
         dynamic: "true",
-        first_published_at: PageViewTracker.getFirstPublishedAt(),
+        first_published_at: getFirstPublishedAt(),
         language: "undefined",
         location: "http://localhost/",
         logged_in_status: "logged in",
@@ -368,7 +320,7 @@ describe("Persisting taxonomy level values", () => {
         taxonomy_level4: "taxo4",
         taxonomy_level5: "taxo5",
         title: "home",
-        updated_at: PageViewTracker.getUpdatedAt(),
+        updated_at: getUpdatedAt(),
       },
     });
   });
