+++ packages/frontend-analytics/src/analytics/pageViewTracker/pageViewTracker.ts
@@ -3,10 +3,19 @@ import {
   PageViewEventInterface,
 } from "./pageViewTracker.interface";
 import { validateParameter } from "../../utils/validateParameter";
-import { FormErrorTracker } from "../formErrorTracker/formErrorTracker";
+import { trackFormError } from "../formErrorTracker/formErrorTracker";
 import { OptionsInterface } from "../core/core.interface";
 import { getTaxonomy, setTaxonomies } from "../../utils/taxonomyUtils";
 import { pushToDataLayer } from "../../utils/pushToDataLayer";
+import {
+  getLanguage,
+  getLocation,
+  getReferrer,
+  getLoggedInStatus,
+  getFirstPublishedAt,
+  getUpdatedAt,
+  getRelyingParty,
+} from "./utils/pageViewTracker.utils";
 
 export class PageViewTracker {
   eventName: string = "page_view_ga4";
@@ -39,8 +48,7 @@ export class PageViewTracker {
 
     if (errorTrigger.length) {
       if (this.enableFormErrorTracking) {
-        const formErrorTracker = new FormErrorTracker();
-        formErrorTracker.trackFormError();
+        trackFormError();
       }
       return;
     }
@@ -54,11 +62,11 @@ export class PageViewTracker {
     const pageViewTrackerEvent: PageViewEventInterface = {
       event: this.eventName,
       page_view: {
-        language: PageViewTracker.getLanguage(),
-        location: PageViewTracker.getLocation(),
+        language: getLanguage(),
+        location: getLocation(),
         organisations: this.organisations,
         primary_publishing_organisation: this.primary_publishing_organisation,
-        referrer: PageViewTracker.getReferrer(),
+        referrer: getReferrer(),
         status_code: validateParameter(parameters.statusCode.toString(), 3),
         title: validateParameter(parameters.englishPageTitle, 300),
         taxonomy_level1: validateParameter(parameters.taxonomy_level1, 100),
@@ -67,13 +75,11 @@ export class PageViewTracker {
           100,
         ),
         content_id: validateParameter(parameters.content_id, 100),
-        logged_in_status: PageViewTracker.getLoggedInStatus(
-          parameters.logged_in_status,
-        ),
+        logged_in_status: getLoggedInStatus(parameters.logged_in_status),
         dynamic: parameters.dynamic.toString(),
-        first_published_at: PageViewTracker.getFirstPublishedAt(),
-        updated_at: PageViewTracker.getUpdatedAt(),
-        relying_party: PageViewTracker.getRelyingParty(),
+        first_published_at: getFirstPublishedAt(),
+        updated_at: getUpdatedAt(),
+        relying_party: getRelyingParty(),
         taxonomy_level3: validateParameter(
           getTaxonomy(parameters.taxonomy_level3, "Level3"),
           100,
@@ -90,84 +96,5 @@ export class PageViewTracker {
     };
 
     pushToDataLayer(pageViewTrackerEvent);
-  }
-
-  /**
-   * Returns a string representing the logged in status.
-   *
-   * @param {boolean} loggedInStatus - The logged in status.
-   * @return {string} The string representation of the logged in status.
-   */
-  static getLoggedInStatus(loggedInStatus: boolean | undefined): string {
-    if (loggedInStatus === undefined) {
-      return "undefined";
-    }
-
-    return loggedInStatus ? "logged in" : "logged out";
-  }
-
-  /**
-   * Returns the value of the 'govuk:first-published-at' meta tag attribute, if it exists.
-   *
-   * @return {string} The value of the 'govuk:first-published-at' meta tag attribute, or "undefined" if it does not exist.
-   */
-  static getFirstPublishedAt(): string {
-    return (
-      document
-        .querySelector('meta[name="govuk:first-published-at"]')
-        ?.getAttribute("content") ?? "undefined"
-    );
-  }
-
-  /**
-   * Retrieves the value of the 'govuk:updated-at' meta tag attribute from the document.
-   *
-   * @return {string} The value of the 'govuk:updated-at' meta tag attribute, or "undefined" if it is not found.
-   */
-  static getUpdatedAt(): string {
-    return (
-      document
-        .querySelector('meta[name="govuk:updated-at"]')
-        ?.getAttribute("content") ?? "undefined"
-    );
-  }
-
-  /**
-   * Returns the hostname of the current document location.
-   *
-   * @return {string} The hostname of the current document location.
-   */
-  static getRelyingParty(): string {
-    return document.location.hostname;
-  }
-
-  /**
-   * Retrieves the language code from the HTML document and returns it in lowercase.
-   *
-   * @return {string} The language code. Defaults to "en" if no language code is found.
-   */
-  static getLanguage(): string {
-    const languageCode = document.querySelector("html")?.getAttribute("lang");
-    return languageCode?.toLowerCase() ?? "undefined";
-  }
-
-  /**
-   * Returns the current location URL as a lowercase string.
-   *
-   * @return {string} The current location URL as a lowercase string, or "undefined" if not available.
-   */
-  static getLocation(): string {
-    return document.location.href?.toLowerCase() ?? "undefined";
-  }
-
-  /**
-   * Retrieves the referrer of the current document.
-   *
-   * @return {string} The referrer as a lowercase string, or "undefined" if it is empty.
-   */
-  static getReferrer(): string {
-    return document.referrer.length
-      ? document.referrer?.toLowerCase()
-      : "undefined";
   }
 }
